demux: libmp4: use the same type for i_depend_on_ES_ID and i_ES_ID
They are defined as bit(16) in ISO/IEC 14496-1:2004.

+++ modules/demux/mp4/libmp4.h
@@ -1049,7 +1049,7 @@ typedef struct MP4_descriptor_ES_s /* ISO/IEC 14496-1 8.3.3 ES_DescrTag */
     int      b_OCRstream;
     int      i_stream_priority;
 
-    int      i_depend_on_ES_ID; /* if b_stream_dependence set */
+    uint16_t  i_depend_on_ES_ID; /* if b_stream_dependence set */
 
     unsigned char *psz_URL;
